refactor(assistant): de-mux non-memory job handlers into their domains
