+++ assistant/src/__tests__/prune-old-conversations-job.test.ts
@@ -9,9 +9,9 @@ mock.module("../util/logger.js", () => ({
 }));
 
 import type { AssistantConfig } from "../config/schema.js";
-import { pruneOldConversationsJob } from "../memory/job-handlers/cleanup.js";
 import { getDb } from "../persistence/db-connection.js";
 import { initializeDb } from "../persistence/db-init.js";
+import { pruneOldConversationsJob } from "../persistence/job-handlers/cleanup.js";
 import type { MemoryJob } from "../persistence/jobs-store.js";
 import {
   conversations,
