+++ assistant/src/media/job-handlers/media-processing.ts
