+++ assistant/src/persistence/job-handlers/cleanup.ts
